restructured the supabase handling as well to differ between auth and db management
