+++ middleware.ts
@@ -1,5 +1,5 @@
 import { type NextRequest } from 'next/server';
-import { updateSession } from './src/lib/supabase/middleware';
+import { updateSession } from './src/lib/supabase/auth/middleware';
 
 export async function middleware(request: NextRequest) {
   return await updateSession(request);
